update unit test files for multiple test suites - improve runtime
